Add files via upload
This is a sample code to demonstrate statistical approach to finding the value of Pi. The code generates a bunch of coordinate pairs within a square, and evaluate whether the point falls in the area of a circle inscribing the square. Using the ratio of points that are in the square to the total number of the squares, we can find the area of the circle, and from it we can find the value of Pi.